Fix win* buildbot warnings: packet_panes.c(304) : error C2220: warning treated as error - no 'object' file generated packet_panes.c(304) : warning C4244: 'function' : conversion from 'gdouble' to 'int', possible loss of data packet_panes.c(304) : warning C4244: 'function' : conversion from 'gdouble' to 'int', possible loss of data
Add casts, old code did the same casts so this should be OK.

svn path=/trunk/; revision=43730

+++ ui/gtk/packet_panes.c
@@ -301,7 +301,7 @@ byte_view_select(GtkWidget *widget, GdkEventButton *event)
     tree_view = GTK_TREE_VIEW(g_object_get_data(G_OBJECT(widget),
                                               E_BYTE_VIEW_TREE_VIEW_PTR));
 
-    byte = bytes_view_byte_from_xy(BYTES_VIEW(widget), event->x, event->y);
+    byte = bytes_view_byte_from_xy(BYTES_VIEW(widget), (gint) event->x, (gint) event->y);
 
     if (byte == -1) {
         return FALSE;
